+++ ship.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  afterPublish ({ exec }) {
+    exec('yarn report-coverage')
+  },
+  testCommandBeforeRelease: ({ isYarn }) => isYarn ? 'yarn test' : 'npm run test'
+}