+++ CHANGELOG.md
@@ -1,8 +1,21 @@
 ## [3.0.1](https://github.com/chimera-js/vue-chimera/compare/v3.0.0...v3.0.1) (2020-03-20)
 
-
+## Fixes
+Fix: passing axios as function
 
 # [3.0.0](https://github.com/chimera-js/vue-chimera/compare/v2.4.3...v3.0.0) (2020-03-18)
 
+## Features
 
+Prefetch reactive endpoints
+
+Mixin strategy and inheritance
+
+SSR compatibility
+
+## Refactors
+
+'Resource' renamed to 'Endpoint'
+
+Refactored and improved reactivity