+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [3.0.1](https://github.com/chimera-js/vue-chimera/compare/v3.0.0...v3.0.1) (2020-03-20)
 
 ## Fixes
-Fix: passing axios as function
+Fix passing axios as function
+Fix send extra params without parameter
 
 # [3.0.0](https://github.com/chimera-js/vue-chimera/compare/v2.4.3...v3.0.0) (2020-03-18)